fix: re-clone corrupt replicas instead of force-recovering them into service
The corruption path started mysqld with innodb_force_recovery and put the
instance back into service. MySQL blocks INSERT, UPDATE and DELETE whenever
innodb_force_recovery is above zero, so that server can neither accept writes
as a primary nor apply relay logs as a replica: it comes up looking healthy
while silently refusing every write and falling permanently behind.

Escalation was also unreachable and untargeted. The marker was cleared on any
successful start, so a level-1 start that came up erased its own state and the
next failure began again at level 1, oscillating between 0 and 1 without ever
reaching 2 or 3 or the exhaustion branch that hands off to auto-reinit. And
escalation ran on any openControl failure, so a boot slower than ReadyTimeout,
a rejected config value or an OOM kill would flip a healthy primary into a
write-rejecting server. The corruption patterns meant to gate it were never
called.

Drop the force-recovery path. Keep the detection, wire it to mysqld's actual
output through a bounded tail retained by FifoLog, and publish the diagnosis in
the Pod termination message so the operator acts on evidence rather than on a
restart count. Match the bare message text case-insensitively: the subsystem
prefix differs between MySQL 8.0's "[InnoDB]" and 5.7/MariaDB's "InnoDB:", so
the old patterns matched neither on 8.0.

Fix the auto-reinit gates. PodFailed alone no longer qualifies: eviction,
preemption and node shutdown all land there with the data volume intact, and
re-cloning those turns a reschedule into a full resync. Only the mysqld
container counts, and a diagnosed corruption re-clones at the low threshold
because restarts cannot repair InnoDB data.

Make RepairReplication refuse to run without GTID auto-positioning, which its
correctness silently depended on, and narrow the metadata patterns so healthy
relay-log messages no longer trigger a reset.

Document the operator-facing behaviour and the manual salvage procedure for a
corrupt primary, which stays a human decision.

Also fix the lint failures the branch introduced.

+++ pkg/management/mysql/diskusage/diskusage.go
@@ -43,7 +43,7 @@ func Of(path string) (Usage, error) {
 	if err := syscall.Statfs(path, &st); err != nil {
 		return Usage{}, err
 	}
-	bsize := int64(st.Bsize)
+	bsize := st.Bsize
 	capacity := int64(st.Blocks) * bsize
 	return Usage{
 		UsedBytes:      capacity - int64(st.Bfree)*bsize,
